+++ src/Constant/MESSAGE.js
@@ -3,5 +3,4 @@ export const MESSAGE = {
   ATTEMPT: "시도할 횟수는 몇 회인가요?\n",
   RESULT: "실행 결과",
   END: "최종 우승자 : ",
-  NO_WINNER: "우승자가 없습니다."
 };
