+++ src/Game/getAttempt.js
@@ -1,6 +1,6 @@
 import { MESSAGE } from '../Constant/MESSAGE';
 import { readLineAsync } from '../Utils/readLineAsync';
-import { verifyAttempt } from '../verify';
+import { verifyAttempt } from '../Verify/verifyAttempt';
 
 export const getAttempt = async () => {
   const attempt = await readLineAsync(MESSAGE.ATTEMPT);
